Move parse() to Config::from_slice()
